limited tile count to 256 per bank

+++ src/tile.rs
@@ -3,7 +3,7 @@ use core::ops::{Index, IndexMut};
 
 /// Unique identifier for a tile. Starts at zero when chip is reset.
 #[derive(Debug, Clone, Copy, Eq, PartialOrd, Ord, PartialEq, Hash, Default)]
-pub struct TileID(pub u16);
+pub struct TileID(pub u8);
 
 /// An array of clusters, each holding 8 pixels
 #[derive(Debug, Clone, Hash, PartialEq, Default)]
